+++ src/main/java/com/hakimi/aviation/service/flight/FlightService.java
@@ -1,12 +1,10 @@
 package com.hakimi.aviation.service.flight;
 
 import com.hakimi.aviation.entity.Flight;
-import com.hakimi.aviation.entity.TicketOrder;
 import com.hakimi.aviation.model.request.flight.BookingRequest;
 import com.hakimi.aviation.model.request.flight.FlightSearchRequest;
 import com.hakimi.aviation.model.vo.FlightSearchVO;
 import com.hakimi.aviation.model.vo.TicketOrderVO;
-import jakarta.servlet.http.HttpServletRequest;
 
 import java.util.List;
 
@@ -16,7 +14,7 @@ public interface FlightService {
 
     List<FlightSearchVO> searchFlightWithCache(FlightSearchRequest request);
 
-    TicketOrderVO bookingFlight(BookingRequest request, Integer userId, String userName);
+    TicketOrderVO bookingFlight(BookingRequest request, Long userId, String userName);
 
 
 }
